Drop Django-Angular (Djangular, djng) entirely!
I really enjoyed some aspects of this library. It enabled me to easily:

- style forms with Bootstrap 3 constructs
- override the widgets used in forms to be custom AngularJS directives
- reference any form field's input value directly AngularJS
- get form validation in the frontend for free

However, using the library also had a lot of costs too. Tightly coupling
presentation and business logic is never great, and coupling yourself to
a legacy JavaScript framework is even worse. AngularJS was my first
frontend framework, but it's been 7 years and the AngularJS community
has largely moved on. It's time for me to move on too.

We're now just using native HTML 5 form validation. Honestly, this works
great. It's available in all modern browsers, is accessible, free of
JavaScript dependency, and more.