Simplify the "active WIMPs" query
This method was formerly much more complicated because it had to span
two different databases (`auth_db` and `ws`). We can simplify it by
changing to a single SQL query.

Additionally, this solves a deprecation warning that will be raised once
we switch to Django 3.1. The ability to pass a raw SQL expression to an
`order_by` clause is going away in Django 4.0.