Fix the ability to pay dues while logged in
The recent deletion of `django-angular` (601498d42f9cc4274c7bf16521a90)
broke this form for participants who are already logged in.

The form was missing required fields and thus would not submit.

Why were form inputs empty?
===========================
`django-angular` had a really nice convenience in that it would handle
Django's `initial` form values (translated to an HTML attribute `value`,
ensuring that we bound that initial value to whatever `ng-model` was.

Without Django-Angular, AngularJS will *not* take the string given in
`value` and assign that to `ng-model`. Rather, it will instead take the
initial (empty) `ng-model` and assign that over the input value.

This bug becomes apparent if you render the underlying inputs - they are
all totally blank!

To fix this (while we continue sunsetting our use of AngularJS), we
employ a hack - using `ng-init` to do a one-time initialization of the
desired value. We'll be moving this form to VueJS soon enough.