Fix three fields being emptied on form load
This bug is a different manifestation of the same core issue described
in a previous commit: 81e39576bde46db165c5ea01c91cfd9b13fa3ad1

Any time we manually set `ng-model` in a form then load said form with
an initial value for the affected inputs, AngularJS will clear out the
input. This makes for frustrating UX.

Of course, the plan is to just axe AngularJS altogether, but we need to
at least fix forms that rely on the `ng-model` trick.

We have a few more forms to fix, namely:
- ParticipantForm (phone numbers & affiliation cleared)
- `LotteryPreferencesForm` (car status, number of passengers
- `ApplicationLeaderForm` (recommendations don't load checked)
- `LeaderForm` (form population broken, but maybe not same issue)

Forms that use `ng-model` but are not currently a problem
- `LotteryPairForm` (it just passes the value to an Angular widget)
- `DuesForm` (already fixed)