Delete the (unused) `/chair/leaders/` page
This was a very old page that predated the existence of
`/{activity}/leaders/`.

Now that we have a page which can show all leaders *and* provides this
same form, there's no reason to have this redundant page.

Repurpose the coverage that we previously had on this view to ensure
full coverage on the related view.