Allow opting in to membership reminders
Motivation behind this feature
==============================
Presently, everybody's MITOC membership expires exactly 365 days after
they paid. We added support for renewing up to 40 days before your
membership expires. *However*, MITOC members don't generally notice that
their membership expired until they do something like try to sign up for
a trip or renew gear.

Why bother encouraging renewal?
===============================
There's absolutely no problem with members going a few weeks or months
between paid years of membership: it's not important to the club that
period of membership be continuous.

However, many MITOCers wish to avoid the annoyance of showing up to the
office with an expired membership (or losing eligibility for discounts).
For these folks, it can be nice to be reminded that it's time to renew.

Key things to build to roll this out to everybody
=================================================
This feature isn't quite ready for use by everybody. Things we need to
build include:

1. Support opting in to "send me a reminder" when paying dues
-------------------------------------------------------------
The most important place to ask people "do you want a reminder to
renew?" is when they are paying dues. We should incorporate a checkbox
in the dues form which lets people select to be reminded.

Of course, this checkbox **must** remain unchecked by default. Sending
emails without express consent is a user-hostile pattern.

2. Support opting into reminders when getting discounts
-------------------------------------------------------
It's important to have continuous membership for many discounts, so we
should expose the ability for participants to opt into getting reminders
when they select discounts.

3. Include a one-click "unsubscribe" link
-----------------------------------------
For CAN-SPAM compliance, and just generally being a system that values
users' time/privacy, we should include an "unsubscribe" link that has a
token for immediate opt-out. This isn't strictly speaking necessary,
since we commit to only notifying users *once* per year (if you're done
being a MITOC member, you can just ignore the email and then never hear
from us again).

However, there are valid reasons to build a simple link:
- Compliance.
- Ease of use (users shouldn't *have* to understand our system or read
  the text in the emails - they should just be able to search for
  "unsubscribe" and click the link)
- Users may pay dues again, but not want a reminder next time