Support one-click unsubscribe for email reminders
Everybody hates companies that *require* you to log in just so they'll
stop sending you emails, right? Let's not replicate that pattern.

This commit creates the ability to unsubscribe from automated emails
with a single click.

For now, we only use this for emails reminding members to renew, but we
could also use the system to:

- Let users manage whether or not they get "you're on a trip!" emails
- Let users subscribe/unsubscribe from Mailman lists