Improve recommendation pre-filling behavior
Presently, when members of the WSC are reviewing WS applications (where
all three members have given recommendations), we pre-fill the form with
the recommendation from whichever WSC member is viewing the page. In
order to assign a rating, the WSC must generally:

1. Untick "is recommendation"
2. Delete their notes
3. Edit the rating to whatever consensus is
4. Click submit ("Create rating")

This is not good UX.

For a multi-chair activity, like Winter School, it doesn't really make
sense to pre-fill the viewing chairs recommendation when trying to
assign a rating. For one thing, it's not necessarily true that it's the
viewing user's recommended rating that will be applied. Second, it
doesn't generally make sense to attach the recommendation's notes to the
rating itself.

Other unrelated changes:
========================
- Add a helper for inspecting form contents
- Check the rating & note pre-filling in more scenarios
- Use `perm_utils` for simpler chair assignment