Explain that "willing to rent/drive" means others
These two options have always meant that the participant is willing to
carpool with others, but a number of participants have a very
understandable preference to not travel with others outside their
household/pod/etc.

For the purposes of trip planning as well as fairness in the lottery, we
should have these participants classified as not being drivers.

The ideal solution here is to add a fourth status for those who are
willing to drive *themselves* (and maybe a partner), just not others.
I'll try to add that as soon as possible as a fast follow.