Add a new choice for driving just yourself
Some people's risk preferences or personal situation might not be
conducive to transporting others.

In normal times, MITOC tries to encourage carpooling for its
environment, social, and practical benefits (limited trailhead/cabin
parking, simpler trip logistics, etc).

However, during COVID, many people may not be willing to spend time in a
car with others. Adding a new driving status lets these people say that
they can transport themselves (or their partner), but not other
participants.

This is currently only an informational status - the goal is to make the
responsibilities of being a driver quite clear, while not giving anybody
an unfair advantage in the lottery.