Warn about the "Winter (outside IAP)" program
If it's currently IAP, and a leader tries to make a normal winter hiking
trip, warn them about the consequences of doing this!

In today's lottery, one trip was accidentally given this program, which
meant that the WS lottery did not run with that trip.

We could consider just *requiring* that all "outside IAP" trips are
indeed outside IAP, but I decline to do that yet because:

- There *might* be a legitimate use case for, say, a midweek trip that's
  on a single trip's lottery system and *not* the full, normal WS lottery
- I don't want to block "weekend after WS" trips from going out.
    - Knowing IAP's exact dates every year isn't possible
    - We *could* just forbid "outside IAP" trips in January, not Feb...
- This is fast & easy, and should solve most problems