Display trip type to chairs, lengthen description
By request of the WSC this year, it's really helpful to have the
"Approve upcoming trips" table specify the trip type.

Previously, we'd been hiding this information so that we can instead
display an indicator about whether or not a trip itinerary has been
submitted. We still keep that indicator, but now move it to be in the
"description" column, which arguably makes more sense anyway (an
itinerary tends to go along with any trip's description).