Consistently order between approval table and view
It's a bit weird to show a list of trips, have somebody click into the
first trip, then click "next trip" and end up with a totally different
trip than what's displayed as the second entry.

To avoid this discrepancy, just define the ordering as an attribute on
the model.