Fix datetime-local widgets on Chrome
The default `format` value from `DateTimeInput` is `YYYY-MM-DD HH:MM:SS`,
but the space separator between date and time is crucially *not* part of
the ISO 8601 spec. Firefox is able to happily parse this, but Chrome
will not.

This means that if you edit an existing trip on Chrome, you're forced to
manually enter *new* datetimes, since the inputs will render to blank.

Similarly, if you go to create a new trip, you won't get any of the
helpful default values and you *must* manually specify when signups open
and when they close.