Add belated migration for SoR rename
Back in September, we renamed MITOC Rock Program back into School of
Rock: 4aab6d9b45ac19b0697efa9fed84ca2bb639ea7e

I forgot to commit a migration changing the `choices` on a new trip's
program. This is thankfully a no-op SQL-wise, so it's really no harm
that it's coming in so late.