Fix bug where old leader signups displayed active
In short, the current system will accept a leader's signup, but then
even if that person is *removed* from the collection of a trip's
leaders, they will remain present in the total leader count (notably,
where the leader count itself is inaccurate)!

There are other parts of that issue that remain in need of a solution,
but this at least corrects the error of miscategorizing leader signups.