Fix a 500 from duplicate leader signups
This fixes a "Server Error" reported on
https://github.com/DavidCain/mitoc-trips/issues/64

It's not a fully robust fix, since we're not actually providing solid UX
for leaders who previously were signed up for a trip and now want to
re-join. (Presently, their only real option is to email the trip leader)

At the very least, though, this will take a 500 and coerce it to a 400,
providing an explanation for what's happening.

Future steps to better this situation:
- Stop rendering "Sign up as leader" since it will always error
- Show users that they signed up before
- Provide a one-click button to sign up again with the same notes