Report the participant's name too
The API endpoint `data/other_verified_users` is something of a misnomer.
It is primarily used to report information we can confirm to have for a
given email address (to be consumed by the membership-processing
pipeline).

Accordingly, we can report a name which we have on file for the user.
This will prove useful when trying to record a membership payment (which
we have previously just assumed should match the first & last name of
whoever has made the payment).