Fix bug where leader-supplied notes are ignored
As a leader, you can pass participant notes in the form (and the API
response will include those notes!) - we just don't persist them to the
database, for display on the trip itself.

The root cause of this is that I just never invoked `signup.save()`. In
other words, this method has *never* worked from a notes perspective,
and I never tested it (or received a leader report explaining the bug).

While we're at it, change this JSON route to stop returning HTML for the
equivalent of a 403 error (use HTTP response codes & a JSON body).

Shout-out to Bjorn Poonen for this report.