Upgrade to Python 3.10
Ubuntu 22.04 LTS is out, which uses Python 3.10 by default.

We also no longer need to deploy the mitoc-member service to the EC2
instance which serves this repo, which means we have a nice opportunity
to provision a brand-new EC2 instance without any of the old
`mitoc-member` files/packages lying around (e.g. `certbot`).

In short, this is a nice opportunity to modernize our version of Python
and also get onto a newer Ubuntu base image.

In the future, when we ideally deploy this repository as a containerized
application, we can make use of Python 3.11 and the great performance
gains the 3.11 release provides!

Other related changes
---------------------
- Install Node 16 (12 is EOL, can't be installed on 22.04)
- Match Poetry version in the container to asdf's version