Refactor & use better types on member stats page
We'll soon aim to replace the membership statistics page with something
more robust. The simplest version would be to just swap out the direct
SQL access for an API call where the API endpoint does the exact same
logic.

Assuming we go with that approach, we can add some better type
annotations now to make the contract clear & allow easy swapping of
implementations. Do that now!