Hit the geardb API to fetch single member
Instead of a direct SQL query, we can start to hit the real API instead!

This will impose some small delays, but the present use of the method is
restricted to an async JS widget and a background task that only runs in
Celery. In the future, we can better leverage caches to avoid hammering
the endpoint needlessly (and to handle the service being down).