Let members specify their name when paying dues
This should be the first steps towards fixing lousy UX for new members:
We've been taking the first & last name from payment information and
using that to populate new member records. Sometimes it's wrong.

For 99% of MITOCers, the credit card they pay with is under their name,
(and the name on their credit card matches the name they prefer to go
by), so it makes no difference where we source their name.

However, a small fraction of people have been really confused to see
that their MITOC membership is listed under:

- a parent's name
- a former name of theirs

Desk workers can always edit a user's name upon request, but this is
silly when we could just *ask people what they want to be called*.

Start doing that now.