Add participant count for signups table
(By request from Asha)

We've already been displaying a signup count on any of the trip listings
(mainly to give participants browsing trips a rough idea of how much
interest there is in that trip), but we've missed an easy opportunity to
do the same on a trip's summary page.

This is expected to be helpful to trip leaders who are trying to figure
out roughly how over (or under) subscribed a given trip in the
pre-lottery stage is.

I'm not terribly sure how helpful it is to count non-placed signups (for
when a WS trip has completed its lottery and some participants were not
placed), but we'll see if that causes any problems.