Handle ValueError possible from HIBP middleware
The current version of `pwned-passwords-django` has an issue - if the
HIBP API includes a comma in its "times seen" output, the library will
try to invoke `int()` on the result and throw a `ValueError`.

I have a sneaking suspicion that commas in the "times seen" was only a
temporary issue with the HIBP API itself -- I last saw exceptions 2 days
ago, and some example links on the filed `pwned-passwords-django`
project includes links to a sample URL which is now comma-less in
output: https://github.com/ubernostrum/pwned-passwords-django/issues/35

While we're at it, just swallow *any* type of exception, since we don't
want to break login if this core bit fails.