Enable logging in with Google SSO
This commit enables logging in with SSO through Google, overriding the
default behavior of `django-allauth`, which is to not let any social
providers vouch that a user definitively owns an email address.

We should soon add a "Sign up with Google" button to the sign up page
too, but I want to make sure this works in production before I get too
much further.