Fix signing up with very-common password
This prevents an integer-parsing error from causing 500s when trying to
sign up with a password that's quite common in HIBP (so common that
there are over 999 known pwnages and a comma is included in the count).

I should probably add some unit tests, but I expect this to be a
short-lived workaround, and I've done some end-to-end testing locally.