Display trips' summaries in the table
This fixes a display bug (quietly tolerated for years), where raw
Markdown would be displayed to end users.

We also take the opportunity to simplify a number of places where trips
are rendered. I'd been adjusting the word/character count to be specific
to the context in which the trip was displayed (pixel hacking,
basically. We can instead just display a really simple summary in all
places.