Rephrase WS application questions for clarity
We've had confusion in past years by people who say "no, I'm not taking
WFA" but then forget to mention that they're already a WEMT or
something. =)

Rephrase two questions to make it very clear that we just want to know
if they expect to hold a current certification by the time WS starts.