Add (mostly informational) reimbursement checkbox
Add a simple boolean value meant to help remind leaders of the need to
register trips for reimbursement. We make use of Google's
form-prefilling to reduce duplicate work for the leaders involved
(but take care to not guess at things).

We're careful to say "if you have approved budget" so that random
leaders don't start submitting requests for reimbursement.

Closes #79