Migrate all WS trips to have consistent levels
This Winter School, we'll be exposing a consistent search interface for
all trips. In order to do this, we'll want to start enforcing consistent
rules on how levels are specified.

In order to do **that,** we'll first need to fix all known previous
trips to properly adhere to the ABC & IS style of ratings.

Most trips are just simple punctuation changes, but a few of these trips
just had totally incorrect levels.