Replace the free-field "level" with activity level
We've historically relied on leaders to write the terrain level and/or
activity level of a trip into one free-form field. There has been ample
confusion over the years, and the resulting variance in human-submitted
values makes categorization/searching difficult. A recent migration
tried to normalize response to this free-field field:
1aa2bb9374be3a64c607e3c9e23e7ce647451476

We *could* try to do some validation and cleaning before saving new
trips, but it's considerably easier UX to just provide leaders an option
of 3 trip types. Furthermore, we *already* ask for the primary trip
activity, so this should neatly capture ice and skiing trips:

- 69 trips include "I" in their level, but only 2 aren't "ice climbing"
  These two trips are mostly hiking, with some ice (they are "CI"
  trips):
    - mitoc-trips.mit.edu/trips/157
    - mitoc-trips.mit.edu/trips/374
- All 51 "S" trips have a `trip_type` of "skiing"