Expose link to trip search, refactor menus
Now that trip search is generally available, give a link so that leaders
can explore it.

It has always been weird that the "Leaders" menu is the item you click
into to create a trip, but the "Trips" menu item only shows existing
trips.

Re-group menu items by topic, not by who can access the functionality.

The WSC and activity chairs still have their own menus, but these are a
very small group of people.