Stop rendering flaked feedback as red when blurred
We've taken the policy of blurring feedback by default when viewing a
participant's page (since most leaders never need to know this
information). There's no reason that we should be leaking whether or not
somebody was regarded as a flake.

Accordingly, hide the "Flaked" button and don't color the text red.

Credit to Bjorn Poonen for the suggestion.