Add a backend-only way to show membership
Now that we've switched to a resilient (idempotent & self-retrying)
membership payment processor, almost all dues are ingested by our
queue-based processor and immediately reported to MITOC Trips.

Accordingly, we can trust the backend cache to be accurate in almost
all circumstances.

This solves several bits of user feedback at once:

- Cole pointed out that we link to the "pay dues" page, even while on
  the "pay dues" page already (that's confusing UX)
- Bjorn noted we don't prompt early waiver renewal like we do for dues
- Several participants have noted that we infinitely stay in the
  "querying servers" stage (a known bug)

It also addresses my general design principle that this project should
be as "old school" as possible (use JS only when absolutely necessary).

This is much faster for end users -- in the usual case (a participant's
membership is cached and already current), we can just report what we
have.