Install and configure ruff
This commit includes some trivial changes to make `ruff` rules pass, but
includes an enumeration of other rules which will require some effort to
enable.

This commit includes adding `noqa` directives in places where we're fine
with the violation, as well as in places we'd really like to fix it. For
an example of each category:

- I occasionally make functions to replace callable classes, and use
  class-like naming syntax to make the replacement clear. Naming rules
  can be bypassed.
- I seriously regret embracing Django signals and should move away from
  them. Their resulting no-op imports should be preserved.

`ruff .` will now pass!