Fix row assignment for single participant discount
A while back, I added `strict=True` to the invocation of `zip()`:
a4a24bac2ed7743062eb4dd918b2e54d0843ddef

Doing so uncovered a bug which I've just left around for a while... It
broke the ability to *immediately* add a participant to the discount
list, but I ignored it, since we can instead just rely on the hourly
whole-sheet refresh.

The bug is basically that we tried to apply a row (3 columns long) to
two matching cells, which would put the name under the email column &
the email under the expiry column.

Start indexing at 1 instead to overwrite the entire row.

It turns out that I never actually noticed this bug since we neglected
to write the cells at all. Oops!