Stop rendering all trips on a single page
This page returns a *massive* amount of content to the end user. It was
suitable back when we had a few hundred trips at most. But there is now
simply too many trips to include on one page.

For now, just redirect to the `/trips/` page -- we've long since stopped
serving up links to `/trips/all`.