Display trips in ascending order by trip date
See the GitHub issue for more discussion and context, but this is meant
to handle the fact that some trips (planned well in advance) end up
sitting at the top of the trips page, week after week.

https://github.com/DavidCain/mitoc-trips/issues/92

There are some downsides to ordering trips this way -- newly-created
trips will now display at the *bottom* of a bunch of upcoming trips. I
may plan to address this later (by, for example, allowing sorting based
on the newest trips).