Quickly summarize a trip's stage
Showing the total number of signups has confused participants for the
past nine years. Mainly, the behavior of the Winter School lottery
(where you sign up for many trips, but are ultimately only placed on
one) has been a large source of confusion. For example, how can a trip
with fifty signups and only ten spaces still have room for more?

Apart from Winter School (where counting signups can be helpful to
organizers, to gauge participant interest in various trips), this is not
the most helpful metric. We should really optimize the upcoming trips
page for *participants*.

This is a first stab at resolving some of the UX problems identified in
https://github.com/DavidCain/mitoc-trips/issues/92

Now that we're able to sort based on a trip's stage, participants can
easily see which trips are open & accepting signups.