Upgrade `ruff` to latest
We unpin `ruff`, which is notably going to make pretty much every
`poetry update` fail, but that's fine -- I want to make an effort to
stay on the latest `ruff`.

There are a few new rules that we either don't care about or can't yet
fix. Just ignore them for now.

It's also important for us to set the target version so as to avoid
false positives on the `I` checker and the `FA102` rule. Specifically,
`ruff` needs to know that:

- you'll always run Python 3.9 or newer to use generic type hints
- you'll always run Python 3.10 or newer in order for `zoneinfo` to be
  available in the stdlib. Otherwise ruff assumes `zoneinfo` to be a
  third party dependency, as it would be for Python 3.8.