Ensure return types match declaration
This is a reasonably important rule to have on if we desire full
typechecking benefits.

With the additional use of `WithAnnotations`, we now need a janky
workaround of destroying the cache directory on every failure. There's a
long GitHub issue about the topic, with this being unfortunately the
only known canonical "fix."