Support non-US plates, prepare to drop localflavor
We also narrow the set of license plate "states" down to just the 50
states and DC. While one should support American Samoa, and US military
plates in the *general* case, they should realistically never come up
when we're recording the license plates of cars that are being used to
drive to New Hampshire from Boston.

Indeed, of the many thousands of cars we have in the system already,
*all* fall under the 50 states or Canada.

We have 42 states represented, including DC. States that are missing:

- Arkansas
- Iowa
- Mississippi
- Nebraska
- North Dakota
- South Dakota
- West Virginia
- Wyoming

To truly drop `localflavor`, we'll need to squash migrations again.