Always hide feedback older than 13 months
Using Django object managers to automatically hide old feedback was a
mistake. Instead, just be very explicit about when we hide old feedback.

This fixes a bug where we've accidentally been showing *all* old
feedback. There was test coverage of the custom manager, but (of course)
not of any actual end-to-end usage!