Remove concept of trip-less feedback
Supporting leaders giving feedback to a participant who wasn't even on a
trip of theirs was an old idea that never materialized. Simplify the
schema by fixing that now.

There's also code which has been making the assumption that given
feedback objects *always* have a trip, so this change is consistent with
that code (and will help mypy pass when type-annotating those lines).