Change scrambling algorithm for whitespace
This approach does a better job preserving the display/spacing of
scrambled text -- when a page is unscrambled, it's now much more likely
to display in exactly the same way.

Problem with the existing approach
==================================
Our naive scrambling of *all* characters has a small problem -- it can
create leading/trailing spaces in the feedback string (which are not
rendered by HTML). It can also create double spaces between a word!

For example:

```python
>>> text = "Hi there, how do you do?"
>>> random.seed(42)
>>> "".join(random.sample(text, len(text)))
' tH,euho  rdwi doyh ?oeo'
```

There's now a leading space *and* a separation of two words.
Rendered text size changes!

```diff
-Hi there, how do you do?
+tH,euho rdwi doyh ?oeo
```

Why not a simpler scramble?
===========================
A simple solution to this is to just scramble the word ordering and the
words themselves. Only problem is that many English words are really
obvious to discern with this strategy, even out of context. A
particularly prying leader could probably deduce a lot of the feedback!

```python
words = text.split()
random.shuffle(words)
return " ".join("".join(random.sample(word, len(word))) for word in words)
```

With seed `42`, this creates:

```diff
-Hi there, how do you do?
+do htee,r hwo yuo iH o?d'
```

Identify permutations for each scrambled word, compare permutations to a
dictionary, and you get:

- `do`, `od` -> `do`
- `htee,r` -> "there", "three", or "ether"
- `how` -> "who" or "how"
- `Hi` -> "hi"

Pretty easy to reverse engineer! It's only more obvious with large
words.