Handle merge of users who probably lost edu email
We've been neglecting to merge users who are obviously the same people
on account of the fact that they likely don't have access to their MIT
email (or BU, Harvard, etc.)

Let's at least put those accounts towards the end of the list. We may
consider just removing them entirely in the future.