Write a dead-simple cache for membership stats
Hitting the gear database with a blocking query is both slow and means
that the trips site is vulnerable to occasional gear database outages
(we also don't want to overwhelm the gear database during periods of
high load, e.g. the first week of WS).

Let's make a super simple cache!