Ensure that waivers use *full* name
It's hopefully obvious that one should use their legal name when signing
legal documents, but nonetheless it's still possible to submit a name
without any spaces in it. If you do that, then the naive attempt to
split a name into first & last name (which we know isn't great) will
fail and the submitted waiver won't be processed by the geardb!

This prevents error responses like these:

```json
{"lastname": ["This field may not be blank."]}
```