Require login to view old trips
As mentioned in a prior commit, there's a scraping script trying to
parse all trips. That's fine, and hopefully it's a one-off script:
33c1063ede2470fa5471649b78f5a6d0254cea0f

However, this is the slowest endpoint served by the application, and we
should be careful to avoid an easy DOS opportunity.

Also, get better about the "Previous trips" label -- when given a very
old date, there's no sense in us showing a "Previous trips" label, so
just dynamically hide it. This also helps get rid of the temporary hack
for hard-coded old dates.