Refer to annual payments (dues) not membership
SOLE wants us to be more careful about who we term as "members" -- as it
is, MITOC's current membership definition excludes people who pay dues
to the club but then only use the club for renting gear.

We can add some clarity to code/comments/documentation by referring to
payment of dues and *membership* as distinct concepts.

I did my best to update as much user-facing language as possible (as
well as code comments), but `git grep -i member` still has a *lot* of
places. We may need to revisit this later.