Fix ordering of waitlist in API route
Bjorn pointed out a weird side effect where after setting `manual_order`
on a waitlist, *new* signups (which will have a null manual order) can
end up earlier on the waiting list, when they should instead appear at
the end of the waitlist. We've been properly handling that in the
Django-based views into waitlist ordering, but we've been reporting a
totally different ordering for the API-based view of trips.

In short, this happened because we were expecting the waitlist models to
define their own ordering, but we explicitly use the *preferred*
ordering elsewhere.

Also, let's fix the API route to start returning JSON responses when the
user lacks permissions.