Fix access control edge case
When specifying a single required group name as a string, we mistakenly
iterate each character of the string as a valid group name!

If groups were named a certain way, this could have made for an access
issue. Oops. Thankfully, there was never an issue.

Use stricter typing too, which would have caught this.