Add ability to duplicate existing trips
This should hopefully prove useful for people trying to use old trips as
the template for new ones.

For now, this remains an undocumented feature -- I'll follow up by
making "duplicate trip" a button available to trip leaders on any trip
(but I'll want to refactor the `trip_edit_buttons` template as part of
doing that).