Declare bankruptcy on frontend CI setup
I haven't touched this in so long that it's likely going to take more
work to fix it than it is to just rebuild from scratch.

Python unit tests are now completely decoupled from any node things,
which they probably should have been in the first place.

The goal now is to modernize the backend setup so that we can hopefully
actually make progress on tearing down the mega-outdated frontend.