Bulk update dependencies
I broke out all other updates that required some build/source code
changes. This batch should be reasonably safe!

I'm going to whittle down the collection of modules for which we ignore
missing imports -- some modules do indeed still not export types, but
we don't actually import them, so it's not a problem.