Upgrade to Python 3.12
Python 3.10 reaches its end of life in October of 2026. Let's get ahead
of that eventual end of support and upgrade to Python 3.12 now!

Switching to `uv` first (and letting uv manage Python versions) made
this *considerably* easier to do.

After making the Django 5.2 upgrade, I'll probably go straight to Python
3.14. We'll also want to upgrade Postgres too, which is long overdue for
an upgrade (though we need to warn the rest of the webmaster team before
doing so in RDS).