Render the aspirational rating on applications
This field was only added for the 2026 application, so let's not render
it on previous years.

The field is currently optional, so that does also mean that we won't
render it on applications where it *was* asked of participants, they
just said nothing. Hopefully that's fine (if not, we could always just
check the year on the application).