Bring back a simple, db-backed leaderboard
The old version was deleted because it heavily relied on AngularJS and
was the only part of the system to require D3. While D3 was quite neat,
a plain and simple table is great too. People can visualize the data as
they see fit, and some simple filters let people be specific in what
they need!

Other uses for this leaderboard include the annual check of which
leaders have led at least two trips, as well as reporting active leaders
to MIT.