Give warning about cross-registered students
With Winter School coming up, MITOC is soon to get a large number of new
members. Let's give a warning that cross-registered students do not
count as MIT students for the purposes of MITOC membership.

What type of membership one should purchase is a continued source of
confusion for many people. While we *could* write a full help page,
hopefully giving a simple warning here covers the most common case.

We could *always* surface the warning directly on `membership.html`, as
Daniel Yang initially did on the (now-closed) pull request:
https://github.com/DavidCain/mitoc-trips/pull/110/

However, we only need to do that on the case of an anonymous viewer,
since the usual flow (used by most MITOCers) is to create an account on
the trips site and *then* use that to more easily pay dues and fill out
a waiver.