Close alum.mit.edu "student" loophole
Technically, it's been possible to cheat the system by claiming to be a
current MIT student but providing your MIT alumni email. Only one person
who's current on dues appears to have done this (and for all we know,
this may have been unintentional). In any case, we may as well prevent
others from trying to use their alum email to unfairly pay lower dues.

We can't guard against all possible bad actors, but it's upsetting that
we need to do this...