Fix query performance for trip search
Locally, searching for "winter school" takes a whopping **six seconds**.

This is because it returns a number of the WS lectures, which have a
very large number of signups to count, and the query plan is not very
efficient.

In production, this led to a system error. There are a couple ways we
could resolve this poorly-performing query. We could *first* obtain the
100 or fewer trip IDs, then query anew. That would fix the issue with
slow `RIGHT JOIN`s to count signups. However, it would discard the
ranking inherent in the query. However, there's a simpler/easier option
-- just stop showing how many spaces remain when you search for a trip.

Given that we only expose this form to leaders *and* most returned
results will be for older trips, "how many spaces remain" is not the
most useful information. If somebody really wants to know, they can
always click into any trip.

The new query plan is slowest at the portion we'd expect -- the actual
bitmap heat scan for the search operation itself (the vectorized search
term being applied to the actual `ws_trip` table).