Improve query for outstanding application count
It's needlessly inefficient to count the recommendations per
application, apply a `DISTINCT`, apple an `ORDER BY`, etc. when all we
really need is to *quickly* count applications that have zero ratings.

Expose a simple method for doing that. On my laptop, this drives query
time from 99ms down to 11ms (when querying 1400 total WS leader
applications).